// openmetadata-ui/src/main/resources/ui/src/enums/GlossaryPage.enum.ts
/*
 *  Copyright 2024 Collate.
 *  Licensed under the Apache License, Version 2.0 (the "License");
 *  you may not use this file except in compliance with the License.
 *  You may obtain a copy of the License at
 *  http://www.apache.org/licenses/LICENSE-2.0
 *  Unless required by applicable law or agreed to in writing, software
 *  distributed under the License is distributed on an "AS IS" BASIS,
 *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 *  See the License for the specific language governing permissions and
 *  limitations under the License.
 */
export enum GlossaryTabs {
  TERMS = 'terms',
  ACTIVITY_FEED = 'activity_feed',
  OVERVIEW = 'overview',
  GLOSSARY_TERMS = 'glossary_terms',
  ASSETS = 'assets',
  CUSTOM_PROPERTIES = 'custom_properties',
}